DatStateMessage.yml and corresponding pydantic_model added
