+++ dat_core/pydantic_models/__init__.py
@@ -1,31 +1,30 @@
-# from dat_core.pydantic_models.configured_dat_catalog import ConfiguredDatCatalog
 from dat_core.pydantic_models.connector_specification import ConnectorSpecification
 from dat_core.pydantic_models.dat_catalog import DatCatalog
 from dat_core.pydantic_models.dat_document_stream import DatDocumentStream, ReadSyncMode, WriteSyncMode
 from dat_core.pydantic_models.dat_log_message import DatLogMessage, Level
-from dat_core.pydantic_models.dat_message import DatMessage, DatDocumentMessage, Data, Type, DatStateMessage, StreamState, StreamStatus
+from dat_core.pydantic_models.dat_message import DatMessage, DatDocumentMessage, Data, Type
 from dat_core.pydantic_models.stream_metadata import StreamMetadata
 from dat_core.pydantic_models.connection import Connection
-# from dat_core.pydantic_models.configured_document_stream import ConfiguredDocumentStream
 from dat_core.pydantic_models.dat_connection_status import DatConnectionStatus, Status
+from dat_core.pydantic_models.dat_state_message import DatStateMessage, StreamState, StreamStatus
 
-__all__ = [
-    'ConnectorSpecification',
-    'DatCatalog',
-    'DatDocumentStream',
-    'ReadSyncMode',
-    'WriteSyncMode',
-    'DatLogMessage',
-    'Level',
-    'DatMessage',
-    'DatDocumentMessage',
-    'Data',
-    'Type',
-    'DatStateMessage',
-    'StreamState',
-    'StreamStatus',
-    'StreamMetadata',
-    'Connection',
-    'DatConnectionStatus',
-    'Status',
-]
+# __all__ = [
+#     'ConnectorSpecification',
+#     'DatCatalog',
+#     'DatDocumentStream',
+#     'ReadSyncMode',
+#     'WriteSyncMode',
+#     'DatLogMessage',
+#     'Level',
+#     'DatMessage',
+#     'DatDocumentMessage',
+#     'Data',
+#     'Type',
+#     'DatStateMessage',
+#     'StreamState',
+#     'StreamStatus',
+#     'StreamMetadata',
+#     'Connection',
+#     'DatConnectionStatus',
+#     'Status',
+# ]
