+++ dat_core/pydantic_models/dat_log_message.py
@@ -1,6 +1,6 @@
 # generated by datamodel-codegen:
 #   filename:  DatLogMessage.yml
-#   timestamp: 2023-12-01T11:07:30+00:00
+#   timestamp: 2024-04-03T10:02:34+00:00
 
 from __future__ import annotations
 
@@ -23,8 +23,8 @@ class DatLogMessage(BaseModel):
     class Config:
         extra = 'allow'
 
-    level: Optional[Level] = Field(None, description='log level of the log message')
-    message: Optional[str] = Field(None, description='log message')
+    level: Level = Field(..., description='log level of the log message')
+    message: str = Field(..., description='log message')
     stack_trace: Optional[str] = Field(
         None,
         description='an optional stack trace if the log message corresponds to an exception',
