+++ dat_core/pydantic_models/dat_state_message.py
@@ -1,0 +1,35 @@
+# generated by datamodel-codegen:
+#   filename:  DatStateMessage.yml
+#   timestamp: 2024-04-03T10:05:33+00:00
+
+from __future__ import annotations
+
+from enum import Enum
+from typing import Any, Dict, Optional
+
+from pydantic import BaseModel, Field
+from dat_core.pydantic_models.dat_document_stream import DatDocumentStream
+
+
+class StreamStatus(Enum):
+    STARTED = 'STARTED'
+    RUNNING = 'RUNNING'
+    COMPLETED = 'COMPLETED'
+
+
+class StreamState(BaseModel):
+    class Config:
+        extra = 'allow'
+
+    data: Dict[str, Any] = Field(..., description='the state data')
+    stream_status: Optional[StreamStatus] = Field(
+        None, description='the stream status')
+
+
+
+class DatStateMessage(BaseModel):
+    class Config:
+        extra = 'allow'
+
+    stream: DatDocumentStream
+    stream_state: StreamState
